Rename traces-observer-service to agent-manager-observer
Rename the directory, Go module path, binary, and env vars
(TRACES_OBSERVER_PORT -> AM_OBSERVER_PORT, OBSERVER_BASE_URL ->
OPENCHOREO_OBSERVER_URL) ahead of consolidating observability into
this service. No behavior changes.
