+++ agent-manager-observer/.air.toml
@@ -4,8 +4,8 @@ tmp_dir = "tmp"
 
 [build]
   args_bin = []
-  bin = "./tmp/amp-traces-observer"
-  cmd = "go build -o ./tmp/amp-traces-observer ."
+  bin = "./tmp/agent-manager-observer"
+  cmd = "go build -o ./tmp/agent-manager-observer ."
   delay = 1000
   exclude_dir = ["assets", "tmp", "vendor", "testdata"]
   exclude_file = []
