+++ agent-manager-observer/.env.example
@@ -1,12 +1,12 @@
 # Server Configuration
-TRACES_OBSERVER_PORT=9098
+AM_OBSERVER_PORT=9098
 
 LOG_LEVEL=INFO
 
 # -----------------------------------------------------------------------------
 # Observer Service Configuration
 # -----------------------------------------------------------------------------
-OBSERVER_BASE_URL=http://localhost:8085
+OPENCHOREO_OBSERVER_URL=http://localhost:8085
 IDP_TOKEN_URL=http://localhost:8090/oauth2/token
 IDP_CLIENT_ID=amp-api-client
 IDP_CLIENT_SECRET=amp-api-client-secret
