+++ agent-manager-observer/cmd/gen-contract/attribute-map.md
