+++ agent-manager-observer/cmd/gen-contract/contract.go
