+++ agent-manager-observer/cmd/gen-contract/contract_test.go
