+++ agent-manager-observer/cmd/gen-contract/main.go
