+++ agent-manager-observer/go.mod
@@ -1,0 +1,5 @@
+module github.com/wso2/agent-manager/agent-manager-observer
+
+go 1.25.0
+
+require github.com/golang-jwt/jwt/v5 v5.3.1
