+++ agent-manager-observer/middleware/cors.go
