+++ agent-manager-observer/observer/auth.go
