+++ agent-manager-observer/observer/client.go
