+++ agent-manager-observer/observer/types.go
