--- traces-observer-service/go.mod
@@ -1,5 +1,0 @@
-module github.com/wso2/agent-manager/traces-observer-service
-
-go 1.25.0
-
-require github.com/golang-jwt/jwt/v5 v5.3.1
